+++ pycmtensor/pycmtensor.py
@@ -3,92 +3,15 @@
 import aesara
 import aesara.tensor as aet
 import biogeme.database as biodb
-import biogeme.expressions as bioexp
 import dill as pickle
 import numpy as np
 import tqdm
-from aesara.tensor.var import TensorVariable
 
+from pycmtensor.expressions import Beta, Weights
 from pycmtensor.functions import errors, full_loglikelihood
 from pycmtensor.utils import learn_rate_tempering
 
 floatX = aesara.config.floatX
-
-
-class Expressions:
-    def __init__(self):
-        pass
-
-    def __add__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return self.sharedVar + other
-        return super().__add__(other)
-
-    def __radd__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return self.sharedVar + other
-        return super().__radd__(other)
-
-    def __sub__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return self.sharedVar - other
-        return super().__sub__(other)
-
-    def __rsub__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return other - self.sharedVar
-        return super().__rsub__(other)
-
-    def __mul__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            if self.sharedVar.ndim > 1:
-                return aet.dot(other, self.sharedVar.T)
-            else:
-                return self.sharedVar * other
-        return super().__mul__(other)
-
-    def __rmul__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            if self.sharedVar.ndim > 1:
-                return aet.dot(other, self.sharedVar.T)
-            else:
-                return other * self.sharedVar
-        return super().__rmul__(other)
-
-    def __div__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return self.sharedVar / other
-        return super().__div__(other)
-
-    def __rdiv__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return other / self.sharedVar
-        return super().__rdiv__(other)
-
-    def __truediv__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return self.sharedVar / other
-        return super().__truediv__(other)
-
-    def __rtruediv__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return other / self.sharedVar
-        return super().__rtruediv__(other)
-
-    def __neg__(self):
-        if isinstance(self, (TensorVariable, Beta)):
-            return -self
-        return super().__neg__()
-
-    def __pow__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return self.sharedVar ** other
-        return super().__pow__(other)
-
-    def __rpow__(self, other):
-        if isinstance(other, (TensorVariable, Beta)):
-            return other ** self.sharedVar
-        return super().__pow__(other)
 
 
 class PyCMTensorModel:
@@ -138,60 +61,6 @@ class PyCMTensorModel:
         return f"{self.name}"
 
 
-class Weights(Expressions):
-    def __init__(self, name, size, status, random_init=True):
-        assert isinstance(size, (list, tuple))
-        rng = np.random.default_rng()
-
-        self.name = name
-        self.size = size
-        self.status = status
-        self.random_init = random_init
-        if len(size) == 1:
-            value = np.zeros(size, dtype=floatX)
-        else:
-            if random_init:
-                value = rng.uniform(
-                    low=-np.sqrt(6.0 / sum(size)),
-                    high=np.sqrt(6.0 / sum(size)),
-                    size=size,
-                )
-            else:
-                value = np.zeros(size, dtype=floatX)
-
-        self.sharedVar = aesara.shared(value=value, name=name)
-        self.shape = self.sharedVar.shape
-        self.sharedVar.__dict__.update({"status": self.status})
-
-    def __call__(self):
-        return self.sharedVar
-
-    def __str__(self):
-        return f"{self.name}"
-
-    def __repr__(self):
-        return f"{self.name}(shape{self.sharedVar.shape.eval()}, TensorSharedVariable)"
-
-    def get_value(self):
-        return self.sharedVar.get_value()
-
-
-class Beta(Expressions, bioexp.Beta):
-    def __init__(self, name, value, lb, ub, status):
-        bioexp.Beta.__init__(self, name, value, lb, ub, status)
-        self.sharedVar = aesara.shared(value=np.array(value, dtype=floatX), name=name)
-        self.sharedVar.__dict__.update({"status": status, "lb": lb, "ub": ub})
-
-    def __call__(self):
-        return self.sharedVar
-
-    def __str__(self):
-        return f"{self.name}"
-
-    def __repr__(self):
-        return f"{self.name}({self.sharedVar.get_value()}, TensorSharedVariable)"
-
-
 class Database(biodb.Database):
     def __init__(self, name, pandasDatabase, choiceVar="CHOICE"):
         super().__init__(name, pandasDatabase)
@@ -202,9 +71,19 @@ class Database(biodb.Database):
             else:
                 variable.x = aet.vector(variable.name)
 
+    def __getstate__(self):
+        return self.__dict__
+
+    def __setstate__(self, state):
+        self.__dict__.update(state)
+
+    def __getattr__(self, name):
+        if name in self.variables:
+            return self.variables[name]
+
     def __getitem__(self, item):
-        """ Returns the aesara.tensor.var.TensorVariable object.
-            Use Database["columnName"] to reference the TensorVariable
+        """Returns the aesara.tensor.var.TensorVariable object.
+        Use Database["columnName"] to reference the TensorVariable
         """
         if hasattr(self.variables[item], "x"):
             return self.variables[item].x
@@ -317,6 +196,10 @@ def build_functions(model, optimizer=None):
 
     model.output_probabilities = aesara.function(
         inputs=model.inputs, outputs=model.p_y_given_x, on_unused_input="ignore",
+    )
+
+    model.output_choices = aesara.function(
+        inputs=model.inputs, outputs=model.pred, on_unused_input="ignore",
     )
 
     model.output_estimated_betas = aesara.function(