+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-## Unreleased
+## v0.3.0 (2022-02-23)
+
+### Feat
+
+- **Database**: add __getattr__ build-in type to Database
+- **pycmtensor.py**: add model.output_choices to generate choices
+
+### Refactor
+
+- moved Beta Weights to expressions.py
 
 ### Fix